small cleanup
Component, Fragments, .jsx, state
